update mock load configuration logic
