+++ JeMPI_Apps/JeMPI_UI/src/services/apiRoutes.ts
@@ -1,5 +1,6 @@
 const ROUTES = {
   POST_FIELDS_CONFIG: '/config',
+  GET_CONFIGURATION: '/configuration',
   POST_CONFIGURATION: '/configuration',
   POST_NOTIFICATIONS: '/notifications',
   POST_INTERACTION: '/interaction',
