+++ JeMPI_Apps/JeMPI_UI/src/test/settings/UniqueToGR.test.tsx
@@ -1,13 +1,13 @@
 
 import { render, waitFor } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import configuration from 'services/configurationData';
+import mockData from 'services/mockData';
 import '@testing-library/jest-dom';
 import UniqueToGR from 'pages/settings/uniqueToGR/UniqueToGR';
 
 
 describe('UniqueToGR', () => {
-  const uniqueGoldenRecordFieldsWithIds = configuration.uniqueGoldenRecordFields.map((row, index) => ({
+  const uniqueGoldenRecordFieldsWithIds = mockData.configuration.uniqueGoldenRecordFields.map((row, index) => ({
     ...row,
     id: `row_${index}`, 
   }));
